// alumni_app/src/app/data/model/empresario.ts
import { Usuario } from "./usuario";

export class Empresario {
    id?: number;
    'estado': boolean;
    'puesto': string;
    'anios': number;
    'usuario': string;
    'email': string;
   
}

export class Empresario2 {
    id?: number;
    'estado': boolean;
    'puesto': string;
    'anios': number;
    'usuario': string;
    'email': string;
   
}